Extract a progress bar widget
