+++ gui/src/widgets/mod.rs
@@ -1,2 +1,3 @@
 pub mod battery;
 pub mod battery_small;
+pub mod progress_bar;
